move utils functions and global state to a separate file
