+++ scripts/openvino.py
@@ -40,46 +40,22 @@ from diffusers import (
     ControlNetModel,
     AutoencoderKL,
 )
+from scripts.global_state_openvino import model_state, pipes
+df_pipe = pipes['diffusers']
+ov_model = pipes['openvino']
 
-current_extension_directory = scripts.basedir() + '/../sd-webui-controlnet'
-print('current_extension_directory', current_extension_directory)
-sys.path.append(current_extension_directory)
+from scripts.utils_ov import ResizeMode, ControlModelType
+controlnet_extension_directory = scripts.basedir() + '/../sd-webui-controlnet'
+sys.path.append(controlnet_extension_directory)
 from scripts.utils import load_state_dict, get_state_dict
-#from internal_controlnet import external_code  # noqa: F403
-#from internal_controlnet.external_code import ControlNetUnit, Preprocessor
 
 #ignore future warnings
 import warnings
 warnings.simplefilter(action='ignore', category=FutureWarning)
 
 
-
-
-class ModelState:
-    def __init__(self):
-        self.enable_ov_extension = False
-        self.enable_caching = False
-        self.recompile = True
-        self.device = "CPU"
-        self.height = 512
-        self.width = 512
-        self.batch_size = 1
-        self.mode = 0
-        self.partition_id = 0
-        self.model_name = ""
-        self.control_models = []
-        self.is_sdxl = False
-        self.lora_model = "None"
-        self.vae_ckpt = "None"
-        self.refiner_ckpt = "None"
-        
-        
-model_state = ModelState()
-
 def cond_stage_key(self):
     return None
-
-
 
 #sdxl invisible-watermark pixel artifact workaround
 class NoWatermark:
@@ -87,95 +63,10 @@ class NoWatermark:
         return img
 
 
-
 def on_change(mode):
     return gr.update(visible=mode)
 
-from modules.processing import StableDiffusionProcessing
-
-POSITIVE_MARK_TOKEN = 1024
-NEGATIVE_MARK_TOKEN = - POSITIVE_MARK_TOKEN
-MARK_EPS = 1e-3
-
-
-def prompt_context_is_marked(x):
-    t = x[..., 0, :]
-    m = torch.abs(t) - POSITIVE_MARK_TOKEN
-    m = torch.mean(torch.abs(m)).detach().cpu().float().numpy()
-    return float(m) < MARK_EPS
-
-
-def mark_prompt_context(x, positive):
-    if isinstance(x, list):
-        for i in range(len(x)):
-            x[i] = mark_prompt_context(x[i], positive)
-        return x
-    if isinstance(x, MulticondLearnedConditioning):
-        x.batch = mark_prompt_context(x.batch, positive)
-        return x
-    if isinstance(x, ComposableScheduledPromptConditioning):
-        x.schedules = mark_prompt_context(x.schedules, positive)
-        return x
-    if isinstance(x, ScheduledPromptConditioning):
-        if isinstance(x.cond, dict):
-            cond = x.cond['crossattn']
-            if prompt_context_is_marked(cond):
-                return x
-            mark = POSITIVE_MARK_TOKEN if positive else NEGATIVE_MARK_TOKEN
-            cond = torch.cat([torch.zeros_like(cond)[:1] + mark, cond], dim=0)
-            return ScheduledPromptConditioning(end_at_step=x.end_at_step, cond=dict(crossattn=cond, vector=x.cond['vector']))
-        else:
-            cond = x.cond
-            if prompt_context_is_marked(cond):
-                return x
-            mark = POSITIVE_MARK_TOKEN if positive else NEGATIVE_MARK_TOKEN
-            cond = torch.cat([torch.zeros_like(cond)[:1] + mark, cond], dim=0)
-            return ScheduledPromptConditioning(end_at_step=x.end_at_step, cond=cond)
-    return x
-
-
-disable_controlnet_prompt_warning = True
-# You can disable this warning using disable_controlnet_prompt_warning.
-
-
-def unmark_prompt_context(x):
-    if not prompt_context_is_marked(x):
-        # ControlNet must know whether a prompt is conditional prompt (positive prompt) or unconditional conditioning prompt (negative prompt).
-        # You can use the hook.py's `mark_prompt_context` to mark the prompts that will be seen by ControlNet.
-        # Let us say XXX is a MulticondLearnedConditioning or a ComposableScheduledPromptConditioning or a ScheduledPromptConditioning or a list of these components,
-        # if XXX is a positive prompt, you should call mark_prompt_context(XXX, positive=True)
-        # if XXX is a negative prompt, you should call mark_prompt_context(XXX, positive=False)
-        # After you mark the prompts, the ControlNet will know which prompt is cond/uncond and works as expected.
-        # After you mark the prompts, the mismatch errors will disappear.
-        if not disable_controlnet_prompt_warning:
-            logger.warning('ControlNet Error: Failed to detect whether an instance is cond or uncond!')
-            logger.warning('ControlNet Error: This is mainly because other extension(s) blocked A1111\'s \"process.sample()\" and deleted ControlNet\'s sample function.')
-            logger.warning('ControlNet Error: ControlNet will shift to a backup backend but the results will be worse than expectation.')
-            logger.warning('Solution (For extension developers): Take a look at ControlNet\' hook.py '
-                  'UnetHook.hook.process_sample and manually call mark_prompt_context to mark cond/uncond prompts.')
-        mark_batch = torch.ones(size=(x.shape[0], 1, 1, 1), dtype=x.dtype, device=x.device)
-        context = x
-        return mark_batch, [], [], context
-    mark = x[:, 0, :]
-    context = x[:, 1:, :]
-    mark = torch.mean(torch.abs(mark - NEGATIVE_MARK_TOKEN), dim=1)
-    mark = (mark > MARK_EPS).float()
-    mark_batch = mark[:, None, None, None].to(x.dtype).to(x.device)
-
-    mark = mark.detach().cpu().numpy().tolist()
-    uc_indices = [i for i, item in enumerate(mark) if item < 0.5]
-    c_indices = [i for i, item in enumerate(mark) if not item < 0.5]
-
-    StableDiffusionProcessing.cached_c = [None, None]
-    StableDiffusionProcessing.cached_uc = [None, None]
-
-    return mark_batch, uc_indices, c_indices, context
-
-df_pipe = None # pipeline from single file of diffusers
-
-OV_df_pipe = None # OpenVINO diffusers pipeline
-
-OV_df_vae = None # vae diffusers model
+from scripts.utils_ov import mark_prompt_context, unmark_prompt_context, POSITIVE_MARK_TOKEN, NEGATIVE_MARK_TOKEN, MARK_EPS, get_control
 
 class OVUnetOption(sd_unet.SdUnetOption):
     def __init__(self, name: str):
@@ -188,230 +79,6 @@ class OVUnetOption(sd_unet.SdUnetOption):
 
 
 from modules.processing import StableDiffusionProcessingImg2Img, StableDiffusionProcessingTxt2Img, StableDiffusionProcessing
-
-def set_numpy_seed(p: processing.StableDiffusionProcessing) -> Optional[int]:
-    """
-    Set the random seed for NumPy based on the provided parameters.
-
-    Args:
-        p (processing.StableDiffusionProcessing): The instance of the StableDiffusionProcessing class.
-
-    Returns:
-        Optional[int]: The computed random seed if successful, or None if an exception occurs.
-
-    This function sets the random seed for NumPy using the seed and subseed values from the given instance of
-    StableDiffusionProcessing. If either seed or subseed is -1, it uses the first value from `all_seeds`.
-    Otherwise, it takes the maximum of the provided seed value and 0.
-
-    The final random seed is computed by adding the seed and subseed values, applying a bitwise AND operation
-    with 0xFFFFFFFF to ensure it fits within a 32-bit integer.
-    """
-    try:
-        # to do: fix multiple seeds in x y z plot or prompt matrix
-        tmp_seed = int(p.all_seeds[0]) # if len(p.seed) == 1 and p.seed == -1 else max((p.seed), 0))
-        tmp_subseed = int(p.all_seeds[0]) # if len(p.subseed) == 1 and p.subseed == -1 else max((p.subseed), 0))
-        seed = (tmp_seed + tmp_subseed) & 0xFFFFFFFF
-        np.random.seed(seed)
-        return seed
-    except Exception as e:
-        logger.warning(e)
-        logger.warning('Warning: Failed to use consistent random seed.')
-        return None
-
-class ControlModelType(Enum):
-    """
-    The type of Control Models (supported or not).
-    """
-
-    ControlNet = "ControlNet, Lvmin Zhang"
-    T2I_Adapter = "T2I_Adapter, Chong Mou"
-    T2I_StyleAdapter = "T2I_StyleAdapter, Chong Mou"
-    T2I_CoAdapter = "T2I_CoAdapter, Chong Mou"
-    MasaCtrl = "MasaCtrl, Mingdeng Cao"
-    GLIGEN = "GLIGEN, Yuheng Li"
-    AttentionInjection = "AttentionInjection, Lvmin Zhang"  # A simple attention injection written by Lvmin
-    StableSR = "StableSR, Jianyi Wang"
-    PromptDiffusion = "PromptDiffusion, Zhendong Wang"
-    ControlLoRA = "ControlLoRA, Wu Hecong"
-    ReVision = "ReVision, Stability"
-    IPAdapter = "IPAdapter, Hu Ye"
-    Controlllite = "Controlllite, Kohya"
-    InstantID = "InstantID, Qixun Wang"
-    SparseCtrl = "SparseCtrl, Yuwei Guo"
-    Unknown = "Unknown"
-
-    @property
-    def is_controlnet(self) -> bool:
-        """Returns whether the control model should be treated as ControlNet."""
-        return self in (
-            ControlModelType.ControlNet,
-            ControlModelType.ControlLoRA,
-            ControlModelType.InstantID,
-        )
-
-    @property
-    def allow_context_sharing(self) -> bool:
-        """Returns whether this control model type allows the same PlugableControlModel
-        object map to multiple ControlNetUnit.
-        Both IPAdapter and Controlllite have unit specific input (clip/image) stored
-        on the model object during inference. Sharing the context means that the input
-        set earlier gets lost.
-        """
-        return self not in (
-            ControlModelType.IPAdapter,
-            ControlModelType.Controlllite,
-        )
-
-    @property
-    def supports_effective_region_mask(self) -> bool:
-        return (
-            self
-            in {
-                ControlModelType.IPAdapter,
-                ControlModelType.T2I_Adapter,
-            }
-            or self.is_controlnet
-        )
-
-class ResizeMode(Enum):
-    """
-    Resize modes for ControlNet input images.
-    """
-
-    RESIZE = "Just Resize"
-    INNER_FIT = "Crop and Resize"
-    OUTER_FIT = "Resize and Fill"
-
-    def int_value(self):
-        if self == ResizeMode.RESIZE:
-            return 0
-        elif self == ResizeMode.INNER_FIT:
-            return 1
-        elif self == ResizeMode.OUTER_FIT:
-            return 2
-        assert False, "NOTREACHED"
-
-def get_control(
-    p: StableDiffusionProcessing,
-    unit, #: ControlNetUnit,
-    idx: int,
-    control_model_type, #: ControlModelType,
-    preprocessor, #: Preprocessor,
-):
-    sys.path.append(current_extension_directory)
-    from internal_controlnet import external_code  # noqa: F403
-    """Get input for a ControlNet unit."""
-    '''
-    if unit.is_animate_diff_batch:
-        unit = add_animate_diff_batch_input(p, unit)
-    '''
-
-    high_res_fix = isinstance(p, StableDiffusionProcessingTxt2Img) and getattr(p, 'enable_hr', False)
-    h, w, hr_y, hr_x = Script.get_target_dimensions(p)
-    input_image, resize_mode = Script.choose_input_image(p, unit, 0)
-    #print('input_image:', input_image) # good
-    if isinstance(input_image, list):
-        print('input image is list')
-        assert unit.accepts_multiple_inputs or unit.is_animate_diff_batch
-        input_images = input_image
-    else: # Following operations are only for single input image.
-        print('input image not list')
-        input_image = Script.try_crop_image_with_a1111_mask(p, unit, input_image, resize_mode)
-        input_image = np.ascontiguousarray(input_image.copy()).copy() # safe numpy
-        if unit.module == 'inpaint_only+lama' and resize_mode == ResizeMode.OUTER_FIT:
-            # inpaint_only+lama is special and required outpaint fix
-            _, input_image = Script.detectmap_proc(input_image, unit.module, resize_mode, hr_y, hr_x)
-        input_images = [input_image]
-        print('image after crop:', input_images)
-
-    if unit.pixel_perfect:
-        unit.processor_res = external_code.pixel_perfect_resolution(
-            input_images[0],
-            target_H=h,
-            target_W=w,
-            resize_mode=resize_mode,
-        )
-    # Preprocessor result may depend on numpy random operations, use the
-    # random seed in `StableDiffusionProcessing` to make the
-    # preprocessor result reproducable.
-    # Currently following preprocessors use numpy random:
-    # - shuffle
-    seed = set_numpy_seed(p)
-    print(f"Use numpy seed {seed}.")
-    print(f"Using preprocessor: {unit.module}")
-    print(f'preprocessor resolution = {unit.processor_res}')
-
-    detected_maps = []
-    def store_detected_map(detected_map, module: str) -> None:
-        if unit.save_detected_map:
-            detected_maps.append((detected_map, module))
-
-    def preprocess_input_image(input_image: np.ndarray):
-        """ Preprocess single input image. """
-        #print('input befre preprocessor cached call:', input_image) # good
-        print('preprocessor:', preprocessor)
-        result = preprocessor.cached_call(
-            input_image,
-            resolution=unit.processor_res,
-            slider_1=unit.threshold_a,
-            slider_2=unit.threshold_b,
-            low_vram=(
-                ("clip" in unit.module or unit.module == "ip-adapter_face_id_plus") and
-                shared.opts.data.get("controlnet_clip_detector_on_cpu", False)
-            ),
-            model=unit.model,
-        )
-        print('result.value:',result.value)
-        print('result.display_images:',result.display_images)
-        detected_map = result.value
-        is_image = preprocessor.returns_image
-        # TODO: Refactor img control detection logic.
-        if high_res_fix:
-            if is_image:
-                hr_control, hr_detected_map = Script.detectmap_proc(detected_map, unit.module, resize_mode, hr_y, hr_x)
-                store_detected_map(hr_detected_map, unit.module)
-            else:
-                hr_control = detected_map
-        else:
-            hr_control = None
-
-        if is_image:
-            print('is_image is true')
-            print('detected map:', detected_map) # None
-            control, detected_map = Script.detectmap_proc(detected_map, unit.module, resize_mode, h, w)
-            print('detected map after:', detected_map) # None
-            print('control after detectmap_proc =', control)
-            store_detected_map(detected_map, unit.module)
-        else:
-            control = detected_map
-            for image in result.display_images:
-                store_detected_map(image, unit.module)
-
-        if control_model_type == ControlModelType.T2I_StyleAdapter:
-            control = control['last_hidden_state']
-
-        if control_model_type == ControlModelType.ReVision:
-            control = control['image_embeds']
-
-        if is_image and unit.is_animate_diff_batch: # AnimateDiff save VRAM
-            control = control.cpu()
-            if hr_control is not None:
-                hr_control = hr_control.cpu()
-
-        return control, hr_control
-
-    def optional_tqdm(iterable, use_tqdm=unit.is_animate_diff_batch):
-        from tqdm import tqdm
-        return tqdm(iterable) if use_tqdm else iterable
-
-    controls, hr_controls = list(zip(*[preprocess_input_image(img) for img in optional_tqdm(input_images)]))
-    assert len(controls) == len(hr_controls)
-    return controls, hr_controls, detected_maps
-
-
-#from diffusers import DiffusionPipeline 
-
-
 
 class OVUnet(sd_unet.SdUnet):
     def __init__(self, p: processing.StableDiffusionProcessing):
@@ -430,7 +97,6 @@ class OVUnet(sd_unet.SdUnet):
         self.has_controlnet = False
     
     def forward(self, x, timesteps=None, context=None, y=None, **kwargs):
-
         print('[OV] forward called, x: min(x), max(x)', torch.min(x), torch.max(x))
         if y is not None:
             print('sd-xl detected')
@@ -523,14 +189,14 @@ class OVUnet(sd_unet.SdUnet):
             print('controlnet detected')
             cond_mark, self.current_uc_indices, self.current_c_indices, context = unmark_prompt_context(context)
 
-            for i in range(len(OV_df_pipe.control_models)):
-                print('control model:', OV_df_pipe.control_models[i])
-                control_model = OV_df_pipe.control_models[i]
-                image = OV_df_pipe.control_images[i]
+            for i in range(len(pipes['openvino'].control_models)):
+                print('control model:', pipes['openvino'].control_models[i])
+                control_model = pipes['openvino'].control_models[i]
+                image = pipes['openvino'].control_images[i]
                 print('image min and max:', torch.min(image), torch.max(image))
                 print('x min and max:', torch.min(x), torch.max(x)) #  tensor(-13.3997) tensor(11.1092)
                 print('context min and max:', torch.min(context), torch.max(context),'shape:', context.shape ) #  tensor(-1024.) tensor(1024.)  [2, 78, 768])
-                control_model = torch.compile(OV_df_pipe.controlnet, backend = 'openvino', options = opt)  # ControlNetModel.from_single_file('./extensions/sd-webui-controlnet/models/control_v11p_sd15_canny_fp16.safetensors', local_files_only=True)
+                control_model = torch.compile(pipes['openvino'].controlnet, backend = 'openvino', options = opt)  # ControlNetModel.from_single_file('./extensions/sd-webui-controlnet/models/control_v11p_sd15_canny_fp16.safetensors', local_files_only=True)
                 down_samples, mid_sample = control_model(
                     x,
                     timesteps,
@@ -565,7 +231,7 @@ class OVUnet(sd_unet.SdUnet):
         else:
             print('no controlnet detected')
         
-        noise_pred = OV_df_pipe.unet(
+        noise_pred = pipes['openvino'].unet(
                 x,
                 timesteps,
                 context,
@@ -593,6 +259,7 @@ class OVUnet(sd_unet.SdUnet):
         self.loaded_config = self.configs[self.profile_idx]
         self.engine.reset(os.path.join(OV_MODEL_DIR, self.loaded_config["filepath"]))
         self.activate(p)
+    
     @staticmethod 
     def prepare_image(
         image,
@@ -606,7 +273,7 @@ class OVUnet(sd_unet.SdUnet):
         guess_mode=False,
     ):  
         from diffusers.image_processor import VaeImageProcessor
-        vae_scale_factor = 2 ** (len(OV_df_pipe.vae.config.block_out_channels) - 1)
+        vae_scale_factor = 2 ** (len(pipes['openvino'].vae.config.block_out_channels) - 1)
         image_processor = VaeImageProcessor(vae_scale_factor=vae_scale_factor, do_convert_rgb=True)
         control_image_processor = VaeImageProcessor(
             vae_scale_factor=vae_scale_factor, do_convert_rgb=True, do_normalize=False
@@ -636,27 +303,21 @@ class OVUnet(sd_unet.SdUnet):
         print('[OV]unload the already loaded unet')
         from modules.sd_models import model_data, send_model_to_trash
         
-        
-        global df_pipe, OV_df_pipe
-
-        
-        
         if model_data.sd_model:
             #send_model_to_trash(model_data.sd_model)
             #model_data.sd_model = None
             model_data.sd_model.state_dict().clear()
             devices.torch_gc()
             print('[OV] finished unloading model')
-        
-        
-        if OV_df_pipe is not None:
-            print('del old OV_df_pipe')
-            del OV_df_pipe
-            del df_pipe
+       
+        if pipes['openvino'] is not None:
+            print('del old pipes[openvino]')
+            #del OV_df_pipe
+            #del df_pipe
             gc.collect()
             print('del old OV_df_pipe and df_pipe')
 
-        OV_df_pipe = OVUnet(p)
+        ov_model = OVUnet(p)
         df_pipe = None
         
         # model state
@@ -678,12 +339,12 @@ class OVUnet(sd_unet.SdUnet):
         else:
             print('load sd-xl pipeline')
             df_pipe = StableDiffusionXLPipeline.from_single_file(checkpoint_path, original_config_file=checkpoint_config, use_safetensors=True, variant="fp32", dtype=torch.float32)
-        OV_df_pipe.unet = df_pipe.unet.to("cpu") # OV device should only be in the options for torch.compile
-        OV_df_pipe.unet = torch.compile(OV_df_pipe.unet, backend="openvino", options = opt)
-        OV_df_pipe.vae = df_pipe.vae.to("cpu")
+        ov_model.unet = df_pipe.unet.to("cpu") # OV device should only be in the options for torch.compile
+        ov_model.unet = torch.compile(ov_model.unet, backend="openvino", options = opt)
+        ov_model.vae = df_pipe.vae.to("cpu")
         print('OpenVINO Extension: loaded unet model')
             
-        OV_df_pipe.has_controlnet = False
+        ov_model.has_controlnet = False
         cn_model="None"
         control_models = []
         control_images = []
@@ -723,7 +384,7 @@ class OVUnet(sd_unet.SdUnet):
                 if control_model_type == ControlModelType.ControlNet:
                     
                     controlnet = ControlNetModel.from_single_file(cn_model_path, local_files_only=False)
-                    OV_df_pipe.controlnet = controlnet
+                    ov_model.controlnet = controlnet
                 else:
                     assert False, f"Control model type {control_model_type} is not supported."
                 
@@ -747,16 +408,17 @@ class OVUnet(sd_unet.SdUnet):
                     controls, hr_controls, additional_maps = get_control(  # get the controlnet input? Yes, preprocess
                         p, unit, 0, ControlModelType.ControlNet, preprocessor)
                     control_images.append(controls[0])
-                
+
+        
         if not control_images: 
-            print('[OV] NO CNET detected')
+            print('[OV] NO CNET detected, unet build finished')
+            pipes['openvino'] = ov_model
             return
 
         print('[OV] cnet detected')
-        OV_df_pipe.has_controlnet = True
-        
-        OV_df_pipe.control_models = control_models
-        OV_df_pipe.control_images = control_images
+        ov_model.has_controlnet = True
+        ov_model.control_models = control_models
+        ov_model.control_images = control_images
         
         print("model_state.control_models:", model_state.control_models)
         
@@ -774,7 +436,7 @@ class OVUnet(sd_unet.SdUnet):
                 elif os.path.isfile(cn_model_path + '.pth'):
                     cn_model_path = cn_model_path + '.pth'
                 controlnet.append(ControlNetModel.from_single_file(cn_model_path, local_files_only=True))
-            OV_df_pipe.controlnet = controlnet
+            ov_model.controlnet = controlnet
         else:
             cn_model_dir_path = os.path.join(scripts.basedir(),'extensions','sd-webui-controlnet','models')
             cn_model_path = os.path.join(cn_model_dir_path, model_state.control_models[0])
@@ -785,132 +447,18 @@ class OVUnet(sd_unet.SdUnet):
             elif os.path.isfile(cn_model_path + '.pth'):
                 cn_model_path = cn_model_path + '.pth'
         
+        pipes['openvino'] = ov_model
         # process image
         print('begin self.prepare_image')
-        for i in range(len(OV_df_pipe.control_images)):
-            OV_df_pipe.control_images[i] = OVUnet.prepare_image(OV_df_pipe.control_images[i], 512, 512, 1, 1, torch.device('cpu'), torch.float32, True, False)
+        for i in range(len(ov_model.control_images)):
+            ov_model.control_images[i] = OVUnet.prepare_image(ov_model.control_images[i], 512, 512, 1, 1, torch.device('cpu'), torch.float32, True, False)
         print('end self.prepare_image')
-        sd_unet.current_unet = OV_df_pipe
+        sd_unet.current_unet = ov_model
         print('end of activate')
         
-
     def deactivate(self):
         print('[OV]deactivate called')
         #del self.unet
-
-
-        
-def align_dim_latent(x: int) -> int:
-    """Align the pixel dimension (w/h) to latent dimension.
-    Stable diffusion 1:8 ratio for latent/pixel, i.e.,
-    1 latent unit == 8 pixel unit."""
-    return (x // 8) * 8
-
-def HWC3(x):
-    assert x.dtype == np.uint8
-    if x.ndim == 2:
-        x = x[:, :, None]
-    assert x.ndim == 3
-    H, W, C = x.shape
-    assert C == 1 or C == 3 or C == 4
-    if C == 3:
-        return x
-    if C == 1:
-        return np.concatenate([x, x, x], axis=2)
-    if C == 4:
-        color = x[:, :, 0:3].astype(np.float32)
-        alpha = x[:, :, 3:4].astype(np.float32) / 255.0
-        y = color * alpha + 255.0 * (1.0 - alpha)
-        y = y.clip(0, 255).astype(np.uint8)
-        return y
-
-    
-def prepare_mask(
-    mask: Image.Image, p: processing.StableDiffusionProcessing
-) -> Image.Image:
-    """
-    Prepare an image mask for the inpainting process.
-
-    This function takes as input a PIL Image object and an instance of the 
-    StableDiffusionProcessing class, and performs the following steps to prepare the mask:
-
-    1. Convert the mask to grayscale (mode "L").
-    2. If the 'inpainting_mask_invert' attribute of the processing instance is True,
-       invert the mask colors.
-    3. If the 'mask_blur' attribute of the processing instance is greater than 0,
-       apply a Gaussian blur to the mask with a radius equal to 'mask_blur'.
-
-    Args:
-        mask (Image.Image): The input mask as a PIL Image object.
-        p (processing.StableDiffusionProcessing): An instance of the StableDiffusionProcessing class 
-                                                   containing the processing parameters.
-
-    Returns:
-        mask (Image.Image): The prepared mask as a PIL Image object.
-    """
-    mask = mask.convert("L")
-    if getattr(p, "inpainting_mask_invert", False):
-        mask = ImageOps.invert(mask)
-
-    if hasattr(p, 'mask_blur_x'):
-        if getattr(p, "mask_blur_x", 0) > 0:
-            np_mask = np.array(mask)
-            kernel_size = 2 * int(2.5 * p.mask_blur_x + 0.5) + 1
-            np_mask = cv2.GaussianBlur(np_mask, (kernel_size, 1), p.mask_blur_x)
-            mask = Image.fromarray(np_mask)
-        if getattr(p, "mask_blur_y", 0) > 0:
-            np_mask = np.array(mask)
-            kernel_size = 2 * int(2.5 * p.mask_blur_y + 0.5) + 1
-            np_mask = cv2.GaussianBlur(np_mask, (1, kernel_size), p.mask_blur_y)
-            mask = Image.fromarray(np_mask)
-    else:
-        if getattr(p, "mask_blur", 0) > 0:
-            mask = mask.filter(ImageFilter.GaussianBlur(p.mask_blur))
-
-    return mask
-
-def get_unique_axis0(data):
-    arr = np.asanyarray(data)
-    idxs = np.lexsort(arr.T)
-    arr = arr[idxs]
-    unique_idxs = np.empty(len(arr), dtype=np.bool_)
-    unique_idxs[:1] = True
-    unique_idxs[1:] = np.any(arr[:-1, :] != arr[1:, :], axis=-1)
-    return arr[unique_idxs]
-
-def nake_nms(x):
-    f1 = np.array([[0, 0, 0], [1, 1, 1], [0, 0, 0]], dtype=np.uint8)
-    f2 = np.array([[0, 1, 0], [0, 1, 0], [0, 1, 0]], dtype=np.uint8)
-    f3 = np.array([[1, 0, 0], [0, 1, 0], [0, 0, 1]], dtype=np.uint8)
-    f4 = np.array([[0, 0, 1], [0, 1, 0], [1, 0, 0]], dtype=np.uint8)
-    y = np.zeros_like(x)
-    for f in [f1, f2, f3, f4]:
-        np.putmask(y, cv2.dilate(x, kernel=f) == x, x)
-    return y
-
-def lvmin_thin(x, prunings=True):
-    y = x
-    for i in range(32):
-        y, is_done = thin_one_time(y, lvmin_kernels)
-        if is_done:
-            break
-    if prunings:
-        y, _ = thin_one_time(y, lvmin_prunings)
-    return y
-
-
-def get_pytorch_control(x: np.ndarray) -> torch.Tensor:
-    # A very safe method to make sure that Apple/Mac works
-    y = x
-
-    # below is very boring but do not change these. If you change these Apple or Mac may fail.
-    y = torch.from_numpy(y)
-    y = y.float() / 255.0
-    y = rearrange(y, 'h w c -> 1 c h w')
-    y = y.clone()
-    y = y.to(devices.get_device_for("controlnet"))
-    y = y.clone()
-    return y
 
 
 opt = dict() # openvino pytorch compile options
@@ -1184,138 +732,6 @@ class Script(scripts.Script):
             detected_map = detected_map[pad_h:pad_h+h, pad_w:pad_w+w]
             detected_map = safe_numpy(detected_map)
             return get_pytorch_control(detected_map), detected_map
-
-    @staticmethod
-    def get_target_dimensions(p: StableDiffusionProcessing) -> Tuple[int, int, int, int]:
-        """Returns (h, w, hr_h, hr_w)."""
-        h = align_dim_latent(p.height)
-        w = align_dim_latent(p.width)
-
-        high_res_fix = (
-            isinstance(p, StableDiffusionProcessingTxt2Img)
-            and getattr(p, 'enable_hr', False)
-        )
-        if high_res_fix:
-            if p.hr_resize_x == 0 and p.hr_resize_y == 0:
-                hr_y = int(p.height * p.hr_scale)
-                hr_x = int(p.width * p.hr_scale)
-            else:
-                hr_y, hr_x = p.hr_resize_y, p.hr_resize_x
-            hr_y = align_dim_latent(hr_y)
-            hr_x = align_dim_latent(hr_x)
-        else:
-            hr_y = h
-            hr_x = w
-
-        return h, w, hr_y, hr_x
-
-    @staticmethod
-    def get_remote_call(p, attribute, default=None, idx=0, strict=False, force=False):
-        if not force and not shared.opts.data.get("control_net_allow_script_control", False):
-            return default
-
-        def get_element(obj, strict=False):
-            if not isinstance(obj, list):
-                return obj if not strict or idx == 0 else None
-            elif idx < len(obj):
-                return obj[idx]
-            else:
-                return None
-
-        attribute_value = get_element(getattr(p, attribute, None), strict)
-        return attribute_value if attribute_value is not None else default
-
-    @staticmethod
-    def choose_input_image(
-            p: processing.StableDiffusionProcessing,
-            unit,#: ControlNetUnit,
-            idx: int
-        ) -> Tuple[np.ndarray, ResizeMode]:
-        """ Choose input image from following sources with descending priority:
-         - p.image_control: [Deprecated] Lagacy way to pass image to controlnet.
-         - p.control_net_input_image: [Deprecated] Lagacy way to pass image to controlnet.
-         - unit.image: ControlNet unit input image.
-         - p.init_images: A1111 img2img input image.
-
-        Returns:
-            - The input image in ndarray form.
-            - The resize mode.
-        """
-        def from_rgba_to_input(img: np.ndarray) -> np.ndarray:
-            if (
-                shared.opts.data.get("controlnet_ignore_noninpaint_mask", False) or
-                (img[:, :, 3] <= 5).all() or
-                (img[:, :, 3] >= 250).all()
-            ):
-                # Take RGB
-                return img[:, :, :3]
-            print("Canvas scribble mode. Using mask scribble as input.")
-            return HWC3(img[:, :, 3])
-
-        # 4 input image sources.
-        p_image_control = getattr(p, "image_control", None)
-        #p_input_image = Script.get_remote_call(p, "control_net_input_image", None, idx)
-        image = unit.get_input_images_rgba()
-        a1111_image = getattr(p, "init_images", [None])[0]
-        print('image, a111-iamge:', image, a1111_image) # good image, None
-
-        resize_mode = unit.resize_mode
-
-        '''
-        if batch_hijack.instance.is_batch and p_image_control is not None:
-            logger.warning("Warn: Using legacy field 'p.image_control'.")
-            input_image = HWC3(np.asarray(p_image_control))
-        elif p_input_image is not None:
-            logger.warning("Warn: Using legacy field 'p.controlnet_input_image'")
-            if isinstance(p_input_image, dict) and "mask" in p_input_image and "image" in p_input_image:
-                color = HWC3(np.asarray(p_input_image['image']))
-                alpha = np.asarray(p_input_image['mask'])[..., None]
-                input_image = np.concatenate([color, alpha], axis=2)
-            else:
-                input_image = HWC3(np.asarray(p_input_image))
-        '''
-        if image is not None:
-            print('unit.get_input_images_rgba() is not none')
-            assert isinstance(image, list)
-            # Inpaint mask or CLIP mask.
-            if unit.is_inpaint or unit.uses_clip:
-                # RGBA
-                input_image = image
-            else:
-                # RGB
-                input_image = [from_rgba_to_input(img) for img in image]
-
-            if len(input_image) == 1:
-                input_image = input_image[0]
-        elif a1111_image is not None:
-            print(' a1111_image is not none')
-            input_image = HWC3(np.asarray(a1111_image))
-            a1111_i2i_resize_mode = getattr(p, "resize_mode", None)
-            assert a1111_i2i_resize_mode is not None
-            resize_mode = external_code.resize_mode_from_value(a1111_i2i_resize_mode)
-
-            a1111_mask_image : Optional[Image.Image] = getattr(p, "image_mask", None)
-            if unit.is_inpaint:
-                if a1111_mask_image is not None:
-                    a1111_mask = np.array(prepare_mask(a1111_mask_image, p))
-                    assert a1111_mask.ndim == 2
-                    assert a1111_mask.shape[0] == input_image.shape[0]
-                    assert a1111_mask.shape[1] == input_image.shape[1]
-                    input_image = np.concatenate([input_image[:, :, 0:3], a1111_mask[:, :, None]], axis=2)
-                else:
-                    input_image = np.concatenate([
-                        input_image[:, :, 0:3],
-                        np.zeros_like(input_image, dtype=np.uint8)[:, :, 0:1],
-                    ], axis=2)
-        else:
-            # No input image detected.
-            #if batch_hijack.instance.is_batch:
-            #    shared.state.interrupted = True
-            raise ValueError("controlnet is enabled but no input image is given")
-
-        assert isinstance(input_image, (np.ndarray, list))
-        return input_image, resize_mode
-
     
     def after_extra_networks_activate(self, p,  *args, **kwargs):
         
@@ -1332,29 +748,23 @@ class Script(scripts.Script):
             scale = float(lora_param.positional[1])
             names.append(name)
             scales.append(scale)
+            
+        if not names: return
         names.sort()
         
-        if OV_df_pipe.lora_fused[''.join(names)] == True: return
-        OV_df_pipe.lora_fused[''.join(names)] = True
+        if pipes['openvino'] and names and pipes['openvino'].lora_fused[''.join(names)] == True: return
+        pipes['openvino'].lora_fused[''.join(names)] = True
 
-        global df_pipe
-            
         for name in names:
             df_pipe.load_lora_weights(os.path.join(os.getcwd(), "models", "Lora"), weight_name=name + ".safetensors", adapter_name=name, low_cpu_mem_usage=True)
 
         df_pipe.set_adapters(names, adapter_weights=scales)
-        
         df_pipe.fuse_lora(adapter_names=names, lora_scale=1.0)
-        
         df_pipe.unload_lora_weights()
     
     def build_unet(self, p):
         OVUnet.activate(p) # reload the engine
         
-        
-        
-        
-
     def process(self, p, *args):
         print("[OV]ov process called")
         model_state.refiner_ckpt = p.refiner_checkpoint
@@ -1386,7 +796,6 @@ class Script(scripts.Script):
         if enable_caching:
             opt_new["cache_dir"] = dir_path
             os.makedirs(dir_path, exist_ok=True)
-            
         else:
             if "cache_dir" in opt_new: del opt_new["cache_dir"]
             if os.path.exists(dir_path) and os.path.isdir(dir_path):
@@ -1411,18 +820,14 @@ class Script(scripts.Script):
             model_state.recompile = True
             print('set recompile to true due to control models change')
             model_state.control_models = current_control_models
-        
-        
-        
         opt = opt_new
-
         if model_state.recompile:
             print('[OV]recompile')
-            try:
-                self.build_unet(p) # rebuild unet from safe tensors
-            except Exception as e:
-                print(f'[OV]Error build_unet: {e}')
-                return
+            #try:
+            self.build_unet(p) # rebuild unet from safe tensors
+            #except Exception as e:
+            #    print(f'[OV]Error build_unet: {e}, fallback to default unet')
+            #    return
         
         self.apply_unet(p) # hook the forward function of unet, do this for every process call
         
@@ -1430,7 +835,6 @@ class Script(scripts.Script):
         # to do: add feature to fallback default vae after replacement
         if loaded_vae_file is not None:
             print('hook the vae')
-            global OV_df_vae
             if OV_df_vae is None: 
                 OV_df_vae = AutoencoderKL.from_single_file(loaded_vae_file, local_files_only=True)
             def vaehook(img):
@@ -1450,20 +854,11 @@ class Script(scripts.Script):
     def apply_unet(self, p):
         if sd_unet.current_unet is not None:
             sd_unet.current_unet.deactivate()
-        
-        
-        
         sd_ldm = p.sd_model
         model = sd_ldm.model.diffusion_model
-
         model._original_forward = model.forward
-        
-        
-        
-        
-        
         print('force forward ')
-        model.forward = OV_df_pipe.forward
+        model.forward = pipes['openvino'].forward
         print('finish force forward ')
     
     def restore_unet(self,p):
@@ -1483,12 +878,12 @@ class Script(scripts.Script):
 
 def refiner_cb_fn(args):
     
-    if model_state.enable_ov_extension and OV_df_pipe and OV_df_pipe.process.extra_generation_params.get('Refiner', False):
+    if model_state.enable_ov_extension and pipes['openvino'] and pipes['openvino'].process.extra_generation_params.get('Refiner', False):
         refiner_filename = model_state.refiner_ckpt.split(' ')[0]
         print('[OV] reload refiner checkpoint:',refiner_filename)
-        OVUnet.activate(OV_df_pipe.process, refiner_filename)
-    elif model_state.enable_ov_extension and OV_df_pipe:
+        OVUnet.activate(pipes['openvino'].process, refiner_filename)
+    elif model_state.enable_ov_extension and pipes['openvino']:
         print("[OV]re-apply main model weights")
-        OVUnet.activate(OV_df_pipe.process)
+        OVUnet.activate(pipes['openvino'].process)
         
 script_callbacks.on_model_loaded(refiner_cb_fn)
